Add iterators to detected object set.
Previously a vector of detections was extracted form the set which
could be inspected. This change adds iterators that can be used to
inspect the set. This approach allows for object sets larger than
what will fit into memory to be handled. Note that the current
implementation has not changed. It is just the API that has changed
to allow for larger sets of detected objects.